Raspi_terminal/requirements.txt
# 路口车流识别系统依赖包
# 适用于Raspberry Pi 5

# 深度学习和计算机视觉
ultralytics==8.0.196          # YOLOv8
opencv-python==4.8.1.78       # OpenCV
numpy==1.24.3                 # 数值计算
Pillow==10.0.0                # 图像处理

# 网络通信
requests==2.31.0              # HTTP请求
urllib3==2.0.4                # URL处理

# 数据处理
pandas==2.0.3                 # 数据分析
json5==0.9.14                 # JSON处理

# 系统监控
psutil==5.9.5                 # 系统信息
GPUtil==1.4.0                 # GPU监控（如果有）

# 日志和配置
loguru==0.7.0                 # 日志管理
pyyaml==6.0.1                 # YAML配置文件
python-dotenv==1.0.0          # 环境变量

# 多线程和异步
threading-timer==0.1.0        # 定时器
asyncio-mqtt==0.13.0          # MQTT异步客户端（可选）

# 树莓派专用
RPi.GPIO==0.7.1               # GPIO控制（仅树莓派）
picamera2==0.3.12             # 树莓派摄像头v2

# 开发和测试
pytest==7.4.0                 # 单元测试
pytest-asyncio==0.21.1        # 异步测试
black==23.7.0                 # 代码格式化
flake8==6.0.0                 # 代码检查

